refactor: rename page components
