+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { FC } from "react";
 import { Routes, Route } from "react-router-dom";
 import { Layout } from "antd";
-import LayoutChoices from "./components/LayoutChoices";
-import ContentBuilder from "./components/ContentBuilder";
-import PagesTable from "./components/PagesTable";
+import LayoutSelector from "./components/LayoutSelector";
+import PageBuilder from "./components/PageBuilder";
+import Home from "./components/Home";
 
 import "./App.css";
 
@@ -11,9 +11,9 @@ const App: FC = () => (
   <div className="App">
     <Layout style={{ height: "100%" }}>
       <Routes>
-        <Route path="/" element={<PagesTable />} />
-        <Route path="layout" element={<LayoutChoices />} />
-        <Route path="content" element={<ContentBuilder />} />
+        <Route path="/" element={<Home />} />
+        <Route path="layout" element={<LayoutSelector />} />
+        <Route path="content" element={<PageBuilder />} />
       </Routes>
     </Layout>
   </div>
