+++ src/components/LayoutSelector/LayoutOptions/Information/index.tsx
